Added DataLayerImport class for DB data insertion
"Introduced the `DataLayerImport` class, which handles database data insertion, connecting to the database and executing INSERT queries. "
