--- DAL/DBInterfaceFacade.java
@@ -1,11 +1,0 @@
-package DAL;
-
-import java.util.List;
-
-public interface DBInterfaceFacade {
-    void addData(String btitle, String a, String yp);
-    boolean checkBook(String bname);
-    void updateBook(String btitle, String ubtitle, String a, String yp);
-    public void deleteBook(String title);
-    public List<String> showAllBooks();
-}
